[UI]: several UI classes renamed and moved to separate files.
Changelog:
- 'UIControls' renamed to 'UI_ControlPanel';
- 'UIDisplay' renamed to 'UI_OutputPanel';
- added 'UI' folder to 'core/classes/';
- several ui files moved to new folder';
- updated ui class import.
